+++ src/app/Http/Controllers/Web/Admin/UserController.php
@@ -25,8 +25,12 @@ class UserController extends Controller
      */
     public function create()
     {
+        $userTypes = auth()->user()->isManager()
+            ? [UserType::Employee => UserType::getDescription(UserType::Employee)]
+            : UserType::asSelectArray();
+
         return view('admin.user.create', [
-            'userTypes' => UserType::asSelectArray(),
+            'userTypes' => $userTypes,
         ]);
     }
 
@@ -35,6 +39,11 @@ class UserController extends Controller
      */
     public function store(StoreUserRequest $request)
     {
+        // Managers can only create Employees
+        if (auth()->user()->isManager() && $request->user_type != UserType::Employee) {
+            abort(403, 'Managers can only create Employee accounts.');
+        }
+
         User::create($request->validated());
 
         alert()->success('User has been added successfully');
@@ -56,9 +65,18 @@ class UserController extends Controller
      */
     public function edit(User $user)
     {
+        // Managers cannot edit Admins or other Managers
+        if (auth()->user()->isManager() && !$user->isEmployee()) {
+            abort(403, 'Managers can only edit Employee accounts.');
+        }
+
+        $userTypes = auth()->user()->isManager()
+            ? [UserType::Employee => UserType::getDescription(UserType::Employee)]
+            : UserType::asSelectArray();
+
         return view('admin.user.edit', [
             'user' => $user,
-            'userTypes' => UserType::asSelectArray(),
+            'userTypes' => $userTypes,
         ]);
     }
 
@@ -82,9 +100,13 @@ class UserController extends Controller
      */
     public function destroy(User $user)
     {
+        if (auth()->user()->isManager()) {
+            abort(403, 'Managers are not allowed to delete users.');
+        }
+
         $user->delete();
 
-        alert()->success('User has been updated successfully');
+        alert()->success('User has been deleted successfully');
         return redirect()->route('admin.user.index');
     }
 }
